tidy up the initiative types and fix issue that retrieves all users to the admin form instead of only the admins
