+++ src/pages/monitoring/outlets/initiativesAdmin/types/initiativeData.ts
@@ -1,45 +1,14 @@
-import type { UserLevel } from "pages/monitoring/types/catalog";
 import type {
-  LocationCompleteInfo,
-  ODataInitiativeShortEntry,
-} from "pages/monitoring/types/odataResponse";
-
-// NOTE: Información a suministrar para crear una iniciativa
-export type GeneralInfo = {
-  name: string;
-  shortName?: string;
-  description: string;
-  objective?: string;
-  influenceArea?: string;
-};
-
-type InitiativeAditionalInfo = {
-  creationDate: string;
-  coordinate: [number, number];
-  polygonArea: number;
-  enabled: boolean;
-  tags: string[];
-};
-
-export type ImagesData = {
-  imageUrl?: File | string | null;
-  bannerUrl?: File | string | null;
-};
-
-export type LocationDataBasic = {
-  locationId: number;
-  locality?: string;
-};
-
-export type InitiativeContact = {
-  phone?: string;
-  email: string;
-};
-
-export type UserData = {
-  userName: string;
-  level: UserLevel;
-};
+  GeneralInfo,
+  LocationDataBasic,
+  InitiativeContact,
+  UserData,
+  ImagesData,
+  ContactSRC,
+  UserSRC,
+  InitiativeAditionalInfo,
+} from "pages/monitoring/types/initiative";
+import type { LocationCompleteInfo } from "pages/monitoring/types/odataResponse";
 
 export type InitiativeDataForm = {
   general: GeneralInfo;
@@ -63,39 +32,10 @@ export type InitiativeDataFormErr = {
   images: ErrorsImages;
 };
 
-// NOTE: Data recibida del Servidor
-export type WithID<T> = T & { id: number };
-type LocationSRC = LocationCompleteInfo;
-type ContactSRC = WithID<InitiativeContact & { initiativeId: number }>;
-type UserSRC = WithID<
-  UserData & { initiativeId: number; creationDate: string }
->;
-
-export type InitiativeFullInfo = WithID<
-  GeneralInfo &
-    ImagesData &
-    InitiativeAditionalInfo & {
-      locations: LocationSRC[];
-      contacts: ContactSRC[];
-      users: UserSRC[];
-    }
->;
-
-export type InitiativeDisplayInfo = Omit<InitiativeFullInfo, "locations"> & {
-  locations: LocationObj[];
-};
-
-export type InitiativeDisplayInfoShort = Omit<
-  ODataInitiativeShortEntry,
-  "locations"
-> & {
-  locations: LocationObj[];
-};
-
 export type CardInfoGrouped = {
   id: number;
   general: GeneralInfo & Omit<InitiativeAditionalInfo, "id">;
-  locations: LocationSRC[];
+  locations: LocationCompleteInfo[];
   contacts: ContactSRC[];
   users: UserSRC[];
   images: ImagesData;
@@ -127,32 +67,3 @@ export type ItemEditorProps<T> = {
   discard?: () => void;
   disabled?: boolean;
 };
-
-// NOTE: Objetos para facilitar la manipulacion de la información
-export type LocationObj = {
-  id: number;
-  departmentId: number;
-  department: string;
-  municipalityId: number | null;
-  municipality: string | null;
-  locality: string | null;
-};
-
-export function isLocationObj(location: unknown): location is LocationObj {
-  return (
-    typeof location === "object" &&
-    location !== null &&
-    "departmentId" in location &&
-    typeof location.departmentId === "number" &&
-    "department" in location &&
-    typeof location.department === "string" &&
-    "municipalityId" in location &&
-    (location.municipalityId === null ||
-      typeof location.municipalityId === "number") &&
-    "municipality" in location &&
-    (location.municipality === null ||
-      typeof location.municipality === "string") &&
-    "locality" in location &&
-    (location.locality === null || typeof location.locality === "string")
-  );
-}
